+++ Locales/enUS.lua
@@ -29,12 +29,15 @@ local L = Shared.Locale
 --L["Guild Bank"] = "Guild Bank"
 --L["Void Storage"] = "Void Storage"
 --L["Mail"] = "Mail"
+--L["Hide Background"] = "Hide Background"
+--L["Hide Slots"] = "Hide Slots"
 
 -- Using short keys for these long strings, so enUS needs to be defined as well
 L["NOTES_BAGS_CLASSIC"] = "This group skins the Backpack, the Keyring, Main Bags, and Bank Bags."
 L["NOTES_BACKPACK"] = "This group skins the Backpack.  If you have enabled the Combined Backpack, it will only skin the slots from the real Backpack and not other bags."
 L["NOTES_MAIN_BAGS"] = "This group skins the main Bags other than the Backpack and Reagent Bag.  If you have enabled the Combined Backpack, it will only skin the slots from those bags and not the Backpack."
 L["NOTES_MAIL"] = "This group skins the Inbox, Send Mail, and Open Mail attachments."
-
 L["OPTIONS_TITLE_MAIN"] = "%s Extended Options"
 L["OPTIONS_DESCRIPTION_MAIN"] = "These options allow you to customize inventory elements to make them look better with certain Masque skins, such as by hiding background artwork."
+L["OPTIONS_DESCRIPTION_SLOTS"] = "Hide the slot artwork for these items."
+L["OPTIONS_DESCRIPTION_GUILDBANK_BACKGROUND"] = "Hide the black background behind Guild Bank items."